requirements.txt
Pillow
pathlib2
Sip
PyQt5
numpy
scipy
matplotlib
lxml